CMakeLists refactor
