+++ tests/core/test_backtest_engine.cpp
