+++ tests/core/test_execution_engine.cpp
