+++ tests/strategies/test_grid_trading.cpp
@@ -13,7 +13,7 @@
 #include <memory>
 
 #include "core/execution_engine/execution_engine.h"
-#include "core/strategy/grid_trading.h"
+#include "core/strategy/grid_trading/grid_trading.h"
 #include "core/backtest_engine/backtest_engine.h"
 #include "utils/logger/log_level.h"
 #include "utils/logger/logger.h"
